# Improve BigQuery crawler performance [sc-6022]