--- a/metaphor/bigquery/README.md
+++ b/metaphor/bigquery/README.md
@@ -41,7 +41,15 @@ See [Common Configurations](../common/README.md) for more information on `output
 
 See [Filter Configurations](../common/docs/filter.md) for more information on the optional `filter` config.
 
-#### Notes
+#### Concurrency
+
+The max number of concurrent requests to the google cloud API can be configured as follows,
+
+```yaml
+max_concurrency: <max_number_of_queries> # Default to 10
+```
+
+### Notes
 
 Make sure to use BigQuery project ID when setting the `database` field in the filter configuration. For example: