--- a/metaphor/bigquery/config.py
+++ b/metaphor/bigquery/config.py
@@ -15,6 +15,9 @@ class BigQueryRunConfig(BaseConfig):
     # Project ID to use. Use the service account's default project if not set
     project_id: Optional[str] = None
 
+    # Max number of concurrent requests to bigquery or logging API, default is 10
+    max_concurrency: Optional[int] = 10
+
     # Include or exclude specific databases/schemas/tables
     filter: Optional[DatasetFilter] = dataclass_field(
         default_factory=lambda: DatasetFilter()