<!-- ☝️ give your PR a short, but descriptive title. -->

### 🤔 Why?

As title.

<!--
  Give reviewers the context necessary to understand this PR. For example,
  a link to the associated Shortcut story, or a few words describing the
  problem this PR solves.

  e.g. [SC000](https://app.shortcut.com/metaphor-data/story/000)
-->

### 🤓 What?

- Call google API in parallel

<!--
  Summary of the changes committed. How does your PR fix the above issue?
-->

### 🧪 Tested?

- Run `metaphor.bigquery.profile` locally. Speed up from 60 sec to 8.8 secs with 10 concurrencies.
- Run `metaphor.bigquery` locally. Speed up from 22 secs to 3.3 secs with 10 concurrencies.

Both MCE files were manually checked, no different.

<!--
  Describe how the change was tested end-to-end.
-->